+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAZCQ5GPXGEYCD71MEJSPW8M"
+        "01KB016ZXPPZDDD7N7JGNEGPS4"
     )
     r"""The unique identifier of the data source"""