+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -130,7 +130,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAZCQ5GKXK1S9MZVHGAX6FW5"
+        "01KB016ZXBVCEVH87CY4N484C7"
     )
     r"""The unique identifier of the data source"""