+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -80,5 +80,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-26T06:13:13.969Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-26T12:11:22.638Z")
     r"""The date and time the resource was last updated"""