+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAZCQ5GNC14ZGGG3TD2FH0VQ"
+        "01KB016ZXD0HDT51GPKNHXMBK4"
     )
     r"""The unique identifier of the data source"""