+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KAZCQ5GRBGYQKZXW8YNXX4B7"
+        "01KB016ZXR0WVJS9RRA4X87CZ5"
     )
     r"""The unique identifier of the data source"""