+++ packages/orq-rc/RELEASES.md
@@ -4229,3 +4229,13 @@ Based on:
 - [python v4.1.0-rc.7] packages/orq-rc
 ### Releases
 - [PyPI v4.1.0-rc.7] https://pypi.org/project/orq-ai-sdk/4.1.0-rc.7 - packages/orq-rc
+
+## 2025-11-26 12:13:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.656.1 (2.753.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.1.0-rc.9] packages/orq-rc
+### Releases
+- [PyPI v4.1.0-rc.9] https://pypi.org/project/orq-ai-sdk/4.1.0-rc.9 - packages/orq-rc